pyproject.toml
[tool.poetry]
name = "drf-api-permit-pg"
version = "0.1.0"
description = ""
authors = ["Thomas Sherer <tcpma2011@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
